Move two more DartTypeUtilities static methods into extensions
