+++ lib/src/extensions.dart
@@ -3,7 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/src/dart/element/member.dart'; // ignore: implementation_imports
 
 extension ElementExtension on Element {
@@ -27,6 +29,95 @@ extension ElementExtension on Element {
       return self;
     }
   }
+}
+
+class EnumLikeClassDescription {
+  final Map<DartObject, Set<FieldElement>> _enumConstants;
+  EnumLikeClassDescription(this._enumConstants);
+
+  /// Returns a fresh map of the class's enum-like constant values.
+  Map<DartObject, Set<FieldElement>> get enumConstants => {..._enumConstants};
+}
+
+extension ClassElementExtension on ClassElement {
+  /// Returns an [EnumLikeClassDescription] for this if the latter is a valid
+  /// "enum-like" class.
+  ///
+  /// An enum-like class must meet the following requirements:
+  ///
+  /// * is concrete,
+  /// * has no public constructors,
+  /// * has no factory constructors,
+  /// * has two or more static const fields with the same type as the class,
+  /// * has no subclasses declared in the defining library.
+  ///
+  /// The returned [EnumLikeClassDescription]'s `enumConstantNames` contains all
+  /// of the static const fields with the same type as the class, with one
+  /// exception; any static const field which is marked `@Deprecated` and is
+  /// equal to another static const field with the same type as the class is not
+  /// included. Such a field is assumed to be deprecated in favor of the field
+  /// with equal value.
+  EnumLikeClassDescription? get asEnumLikeClass {
+    // See discussion: https://github.com/dart-lang/linter/issues/2083.
+
+    // Must be concrete.
+    if (isAbstract) {
+      return null;
+    }
+
+    // With only private non-factory constructors.
+    for (var constructor in constructors) {
+      if (!constructor.isPrivate || constructor.isFactory) {
+        return null;
+      }
+    }
+
+    var type = thisType;
+
+    // And 2 or more static const fields whose type is the enclosing class.
+    var enumConstantCount = 0;
+    var enumConstants = <DartObject, Set<FieldElement>>{};
+    for (var field in fields) {
+      // Ensure static const.
+      if (field.isSynthetic || !field.isConst || !field.isStatic) {
+        continue;
+      }
+      // Check for type equality.
+      if (field.type != type) {
+        continue;
+      }
+      var fieldValue = field.computeConstantValue();
+      if (fieldValue == null) {
+        continue;
+      }
+      enumConstantCount++;
+      enumConstants.putIfAbsent(fieldValue, () => {}).add(field);
+    }
+    if (enumConstantCount < 2) {
+      return null;
+    }
+
+    // And no subclasses in the defining library.
+    if (hasSubclassInDefiningCompilationUnit) return null;
+
+    return EnumLikeClassDescription(enumConstants);
+  }
+
+  bool get hasSubclassInDefiningCompilationUnit {
+    var compilationUnit = library.definingCompilationUnit;
+    for (var cls in compilationUnit.classes) {
+      InterfaceType? classType = cls.thisType;
+      do {
+        classType = classType?.superclass;
+        if (classType == thisType) {
+          return true;
+        }
+      } while (classType != null && !classType.isDartCoreObject);
+    }
+    return false;
+  }
+
+  bool get isEnumLikeClass => asEnumLikeClass != null;
 }
 
 extension AstNodeExtensions on AstNode? {
